+++ smartcontract/cli/src/resource/verify.rs
@@ -1654,7 +1654,7 @@ mod tests {
             last_bgp_up_at: 0,
             last_bgp_reported_at: 0,
             bgp_rtt_ns: 0,
-            feed_pk: Pubkey::default(),
+            feed_pks: vec![],
         }
     }