+++ smartcontract/sdk/go/serviceability/state_test.go
@@ -381,7 +381,7 @@ func TestCustomJSONMarshal(t *testing.T) {
 				"LastBgpReportedAt": 0,
 				"BgpRttNs": 0,
 				"PubKey": "` + dummyPubKeyB58 + `",
-				"FeedPk": "11111111111111111111111111111111"
+				"FeedPks": []
 			}`,
 			expectErr: false,
 		},